+++ raftstore-proxy/ffi/src/RaftStoreProxyFFI/VersionCheck.h
@@ -3,5 +3,5 @@
 #include "@version"
 
 namespace DB {
-constexpr uint32_t RAFT_STORE_PROXY_MAGIC_NUMBER = 0x13579BDF;
+constexpr uint64_t RAFT_STORE_PROXY_MAGIC_NUMBER = 0x13579BDF;
 }  // namespace DB
